refactor identity profile login
