+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Identity/LoginHandlers/IdentityProfileLoginBase.cs
@@ -1,0 +1,37 @@
+﻿using System.Security.Claims;
+using System;
+using Unity.GrantManager.Identity;
+using OpenIddict.Abstractions;
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
+using Volo.Abp.MultiTenancy;
+using Volo.Abp.DependencyInjection;
+using Volo.Abp.Identity;
+using Volo.Abp.PermissionManagement;
+using Microsoft.Extensions.Configuration;
+
+namespace Unity.GrantManager.Web.Identity.LoginHandlers
+{
+    internal abstract class IdentityProfileLoginBase : ITransientDependency
+    {
+        public IAbpLazyServiceProvider LazyServiceProvider { get; set; } = default!;
+        protected ICurrentTenant CurrentTenant => LazyServiceProvider.LazyGetRequiredService<ICurrentTenant>();
+        protected IdentityUserManager IdentityUserManager => LazyServiceProvider.LazyGetRequiredService<IdentityUserManager>();
+        protected IdentityRoleManager IdentityRoleManager => LazyServiceProvider.LazyGetRequiredService<IdentityRoleManager>();
+        protected PermissionManager PermissionManager => LazyServiceProvider.LazyGetRequiredService<PermissionManager>();
+        protected IIdentityUserRepository IdentityUserRepository => LazyServiceProvider.LazyGetRequiredService<IIdentityUserRepository>();
+        protected IConfiguration Configuration => LazyServiceProvider.LazyGetRequiredService<IConfiguration>();
+
+        protected static void AssignDefaultClaims(ClaimsPrincipal claimsPrinicipal, string displayName, Guid userId)
+        {
+            claimsPrinicipal.AddClaim("DisplayName", displayName);
+            claimsPrinicipal.AddClaim("UserId", userId.ToString());
+            claimsPrinicipal.AddClaim("Badge", Utils.CreateUserBadge(displayName));
+        }
+
+        protected static string? GetClaimValue(JwtSecurityToken token, string type)
+        {
+            return token.Claims.FirstOrDefault(s => s.Type == type)?.Value;
+        }
+    }
+}
