+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/appsettings.Development.json
@@ -39,5 +39,8 @@
     "Enrich": [
       "FromLogContext"
     ]
+  },
+  "IdentityProfileLogin": {
+    "AutoCreateUser": true
   }
 }
